Update W10_rickroll
This will set your volume to max, and will not let you change your volume once set.
This will download the MP3, so this may break if the MP3 source is changed. You can change it to whatever MP3 source you want.
This will persist through reboot, so you will need to terminate the service and you should be good to go.
Don't actually use this script.